4е изменение

+++ src/main/cpp/t04_deposit.cpp
@@ -23,7 +23,7 @@ using namespace std;
 
 int t04_deposit() {
 
-double P, X, Y;
+float P, X, Y;
 int K, i=1;
 
 cin >> P;
@@ -36,8 +36,8 @@ P=1+P/100;
  while (i<=K) {
     ++i;
     X=X*P;
-    Y=int(Y*P)+100*(X-int(X));
-    X=int(X+int(Y/100));
+    Y=Y*P+100*(X-int(X));
+    X=int(X)+int(Y/100);
     Y=int(Y)%100;
  }
  cout << X << " " << Y;
